convert key to unsigned short int

+++ BPlusTree/utils.cpp
@@ -5,12 +5,12 @@
 #include "utils.h"
 
 
-std::vector<std::pair<Node*, int> > BPlusTree :: _ancestry(float key) {
+std::vector<std::pair<Node*, int> > BPlusTree :: _ancestry(unsigned short int key) {
 
     std::vector<std::pair<Node*, int> > res;
     Node* curr = root;
     int _, i;
-    float *keys;
+    unsigned short int *keys;
 
     // internal nodes
     for (_=1; _<height; _++) {
@@ -38,14 +38,14 @@ void _shift(LeafNode* node, int src, int dir) {
     int n = (dir==1 ? node->numKeysInserted++ : node->numKeysInserted--) - src;
     if (!n || src+dir<0) return;
     std::memcpy(node->records+src+dir, node->records+src, n*sizeof(Record*));
-    std::memcpy(node->keys+src+dir, node->keys+src, n*sizeof(float)); 
+    std::memcpy(node->keys+src+dir, node->keys+src, n*sizeof(unsigned short int)); 
 }
 
 
 void _shift(InternalNode* node, int src, int dir) {
     int n = (dir==1 ? node->numKeysInserted++ : node->numKeysInserted--) - src;
     if (n>=0) std::memcpy(node->children+src+dir, node->children+src, (n+1)*sizeof(Record*));
-    if (n>0) std::memcpy(node->keys+src+dir, node->keys+src, n*sizeof(float)); 
+    if (n>0) std::memcpy(node->keys+src+dir, node->keys+src, n*sizeof(unsigned short int)); 
 }
 
 
@@ -65,7 +65,7 @@ int _leafSibling(InternalNode* parent, int offset) {
 }
 
 
-void _updateFirstLeft( std::vector<std::pair<Node*, int> > st, float key) {
+void _updateFirstLeft( std::vector<std::pair<Node*, int> > st, unsigned short int key) {
     // update first left parent's key
     for (int i=st.size()-1; i>=0; i--) {
         if (st[i].second > 0) {
@@ -76,7 +76,7 @@ void _updateFirstLeft( std::vector<std::pair<Node*, int> > st, float key) {
 }
 
 
-float _leftmost(Node* node) {
+unsigned short int _leftmost(Node* node) {
     while (dynamic_cast<InternalNode*>(node)) {
         node = ((InternalNode*) node)->children[0];
     }
